perf: gorm默认Id类型搜索

+++ orm/gormv2.go
@@ -117,7 +117,7 @@ func DbFirstBy(out interface{}, where string, args ...interface{}) (err error) {
 }
 
 // DbFirstById 查找
-func DbFirstById(out interface{}, id uint64) error {
+func DbFirstById(out interface{}, id uint) error {
 	return _db.First(out, id).Error
 }
